app-engine/requirements.txt
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.0.1
google-api-core==2.8.0
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-datastore==2.3.0
google-cloud-storage==1.42.2
google-cloud-texttospeech==2.11.0
google-cloud-translate==3.7.3
google-cloud-vision==2.6.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
grpcio==1.46.3
grpcio-status==1.46.3
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==0.4.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.1.1
proto-plus==1.20.4
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
